pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-crm-ai"
version = "0.1.0"
description = "FastAPI mini CRM with async summarization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.2",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "email-validator>=2.1",
  "tenacity>=8.5",
  "aiosqlite>=0.20",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "httpx>=0.27,<0.28",
  "ruff>=0.6",
  "mypy>=1.11",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
  "ignore:'cgi' is deprecated and slated for removal in Python 3.13:DeprecationWarning:httpx.*",
]

[tool.setuptools]
license-files = []

[tool.setuptools.packages.find]
include = ["app*"]
